move liff init to App.vue
